functions/requirements.txt
firebase_functions~=0.1.0
pycryptodome~=3.20.0

